refactor(document processor): Move extract_by_ocr from scrapers.tasks
